backend/requirements.txt
langchain
langchain-core
langchain-openai
langchain-community
langgraph-checkpoint-postgres
psycopg[binary]
dagster
dagster-webserver
pydantic
psycopg2-binary
fastapi
ordered_set